+++ craft_parts/plugins/npm_plugin.py
@@ -20,6 +20,7 @@ import logging
 import os
 import platform
 import re
+from pathlib import Path
 from textwrap import dedent
 from typing import Any, Literal, cast
 
@@ -28,6 +29,7 @@ from pydantic import model_validator
 from typing_extensions import Self, override
 
 from craft_parts.errors import InvalidArchitecture
+from craft_parts.utils.npm_utils import get_install_from_local_tarballs_commands
 
 from . import validator
 from .base import Plugin
@@ -54,12 +56,22 @@ _NODE_ARCH_FROM_PLATFORM = {
 class NpmPluginProperties(PluginProperties, frozen=True):
     """The part properties used by the npm plugin."""
 
-    plugin: Literal["npm"] = "npm"
+    plugin: Literal["npm", "npm-use"] = "npm"
 
     # part properties required by the plugin
     npm_include_node: bool = False
     npm_node_version: str | None = None
     source: str  # pyright: ignore[reportGeneralTypeIssues]
+    build_attributes: list[str] = []
+
+    @model_validator(mode="after")
+    def include_node_not_self_contained(self) -> Self:
+        """If npm-include-node is true, then self-contained must not be in build-attributes."""
+        if self.npm_include_node and "self-contained" in self.build_attributes:
+            raise ValueError(
+                "npm-include-node cannot be true if `self-contained` in build attributes"
+            )
+        return self
 
     @model_validator(mode="after")
     def node_version_defined(self) -> Self:
@@ -134,6 +146,15 @@ class NpmPlugin(Plugin):
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         super().__init__(*args, **kwargs)
         self._node_binary_path: str | None = None
+
+    @property
+    def _is_self_contained(self) -> bool:
+        return "self-contained" in self._part_info.build_attributes
+
+    @property
+    def _npm_cache_backstage(self) -> Path:
+        """Path to npm cache to consume packages from."""
+        return self._part_info.project_info.dirs.backstage_dir / "npm-cache"
 
     @staticmethod
     def _get_architecture() -> str:
@@ -265,6 +286,11 @@ class NpmPlugin(Plugin):
         }
         if cast(NpmPluginProperties, self._options).npm_include_node:
             base_env["PATH"] = "${CRAFT_PART_INSTALL}/bin:${PATH}"
+
+        if self._is_self_contained:
+            # explicitly block registry access during offline builds
+            base_env["npm_config_registry"] = "https://localhost:1"
+
         return base_env
 
     @override
@@ -275,6 +301,9 @@ class NpmPlugin(Plugin):
     @override
     def get_build_commands(self) -> list[str]:
         """Return a list of commands to run during the build step."""
+        if self._is_self_contained:
+            return self._get_self_contained_build_commands()
+
         cmd: list[str] = []
         options = cast(NpmPluginProperties, self._options)
         if options.npm_include_node:
@@ -323,4 +352,24 @@ class NpmPlugin(Plugin):
             """
             )
         ]
+
         return cmd
+
+    def _get_self_contained_build_commands(self) -> list[str]:
+        """Return a list of commands to run during self-contained build step."""
+        pkg_path = self._part_info.part_build_dir / "package.json"
+        bundled_pkg_path = (
+            self._part_info.part_build_subdir / ".parts" / "package.bundled.json"
+        )
+        return [
+            *get_install_from_local_tarballs_commands(
+                pkg_path, bundled_pkg_path, self._npm_cache_backstage
+            ),
+            'npm install --offline -g --prefix "${CRAFT_PART_INSTALL}" "$(npm pack . | tail -1)"',
+        ]
+
+    @classmethod
+    @override
+    def supported_build_attributes(cls) -> set[str]:
+        """Return the build attributes that this plugin supports."""
+        return {"self-contained"}